requirements.txt
# GIGATT Geomapper - Python dependencies
# Install: pip install -r requirements.txt
#
# Note: supabase>=2.26 pulls storage3 with pyiceberg/pyroaring which need C++ build tools.
# Pinning to <2.26 avoids that. Upgrade when pyiceberg has pre-built wheels for your Python.

supabase>=2.0.0,<2.26
PyJWT>=2.8.0
python-dotenv>=1.0.0
pypdf>=3.0.0
pytest>=7.0.0
